I added functionality to the attackButton.
The attackButton, when clicked, calls startAnimation(). I also added the
initial skeleton image to the background when the application begins.
Previously the skeleton image only appeared when the attack button was
clicked. I also changed the loop in animationTimer. Previously, the
animation ran indefinitely after attackButton was clicked. Now it stops
the animation after the sword swings.